etl-files/etl_files_media_server/requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.10,<3.0
requests>=2.31.0
colorlog



